move chat queries to individual files
