+++ emphasis_education_app/src/queries/GetMessages.ts
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export const GET_MESSAGES = gql`
+query getMessages($chatID: String!, $init: Int!) {
+  getMessages(chatID: $chatID, init: $init) {
+    # id
+    _id
+    text
+    createdAt
+    user {
+      _id
+      name
+    }
+  }
+}
+`;
