+++ emphasis_education_app/src/queries/MessageReceived.ts
@@ -1,0 +1,15 @@
+import gql from 'graphql-tag';
+
+export const SUB = gql`
+subscription {
+  messageReceived {
+    text
+    MessageId
+    createdAt
+    user {
+      name
+      _id
+    }
+  }
+}
+`
