+++ emphasis_education_app/src/queries/SendMessage.ts
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export const SEND_MESSAGE = gql`
+  mutation sendMessage($messages: [MessageInput]) {
+    sendMessage(messages: $messages) {
+      # id
+      text
+      MessageId
+      createdAt
+      user {
+        name
+        _id
+      }
+    }
+  }
+`;
